suggestions implemented, clippy warning fix
