+++ crates/subspace-farmer/src/commands.rs
@@ -1,0 +1,3 @@
+mod farm;
+
+pub(crate) use farm::farm;
