+++ crates/subspace-farmer/src/object_mappings.rs
